--- a/Makefile
+++ b/Makefile
@@ -84,8 +84,6 @@ formatting: format-codestyle
 
 .PHONY: test
 test:
-	ls -a
-	ls .pytest_cache
 	poetry run mpiexec -n 4 pytest --cache-clear --with-mpi tests/test_numeric/test_immersed_boundary_ops/
 
 .PHONY: update-dev-deps