--- a/sopht_mpi/numeric/immersed_boundary_ops/EulerianLagrangianGridCommunicatorMPI2D.py
+++ b/sopht_mpi/numeric/immersed_boundary_ops/EulerianLagrangianGridCommunicatorMPI2D.py
@@ -189,94 +189,97 @@ def generate_eulerian_to_lagrangian_grid_interpolation_kernel_2d(
     # grid/problem dimensions
     grid_dim = 2
 
-    @njit(cache=True, fastmath=True)
-    def eulerian_to_lagrangian_grid_interpolation_kernel_2d(
-        lag_grid_field,
-        eul_grid_field,
-        interp_weights,
-        nearest_eul_grid_index_to_lag_grid,
-    ):
-        """Interpolate an Eulerian field onto a Lagrangian field.
-
-        Inputs:
-        the nearest_eul_grid_index_to_lag_grid(grid_dim, num_lag_nodes) and
-        interpolation weights interp_weights of
-        shape (2 * interp_kernel_width, 2 * interp_kernel_width, num_lag_nodes)
-
-        """
-        num_lag_nodes = lag_grid_field.shape[-1]
-        for i in range(0, num_lag_nodes):
-            lag_grid_field[i] = np.sum(
-                eul_grid_field[
-                    nearest_eul_grid_index_to_lag_grid[1, i]
-                    - interp_kernel_width
-                    + 1 : nearest_eul_grid_index_to_lag_grid[1, i]
-                    + interp_kernel_width
-                    + 1,
-                    nearest_eul_grid_index_to_lag_grid[0, i]
-                    - interp_kernel_width
-                    + 1 : nearest_eul_grid_index_to_lag_grid[0, i]
-                    + interp_kernel_width
-                    + 1,
-                ]
-                * interp_weights[..., i]
-            ) * (dx**grid_dim)
-
-    @njit(cache=True, fastmath=True)
-    def vector_field_eulerian_to_lagrangian_grid_interpolation_kernel_2d(
-        lag_grid_field,
-        eul_grid_field,
-        interp_weights,
-        nearest_eul_grid_index_to_lag_grid,
-    ):
-        """Interpolate an Eulerian vector field onto a Lagrangian vector field.
-
-        Inputs:
-        the nearest_eul_grid_index_to_lag_grid(grid_dim, num_lag_nodes) and
-        interpolation weights interp_weights of
-        shape (2 * interp_kernel_width, 2 * interp_kernel_width, num_lag_nodes)
-
-        """
-        num_lag_nodes = lag_grid_field.shape[-1]
-        for i in range(0, num_lag_nodes):
-            # numba doesnt allow multiple axes for np.sum :/,
-            # hence needs to be done serially
-            lag_grid_field[0, i] = np.sum(
-                eul_grid_field[
-                    0,
-                    nearest_eul_grid_index_to_lag_grid[1, i]
-                    - interp_kernel_width
-                    + 1 : nearest_eul_grid_index_to_lag_grid[1, i]
-                    + interp_kernel_width
-                    + 1,
-                    nearest_eul_grid_index_to_lag_grid[0, i]
-                    - interp_kernel_width
-                    + 1 : nearest_eul_grid_index_to_lag_grid[0, i]
-                    + interp_kernel_width
-                    + 1,
-                ]
-                * interp_weights[..., i]
-            ) * (dx**grid_dim)
-            lag_grid_field[1, i] = np.sum(
-                eul_grid_field[
-                    1,
-                    nearest_eul_grid_index_to_lag_grid[1, i]
-                    - interp_kernel_width
-                    + 1 : nearest_eul_grid_index_to_lag_grid[1, i]
-                    + interp_kernel_width
-                    + 1,
-                    nearest_eul_grid_index_to_lag_grid[0, i]
-                    - interp_kernel_width
-                    + 1 : nearest_eul_grid_index_to_lag_grid[0, i]
-                    + interp_kernel_width
-                    + 1,
-                ]
-                * interp_weights[..., i]
-            ) * (dx**grid_dim)
-
     if n_components == 1:
+
+        @njit(cache=True, fastmath=True)
+        def eulerian_to_lagrangian_grid_interpolation_kernel_2d(
+            lag_grid_field,
+            eul_grid_field,
+            interp_weights,
+            nearest_eul_grid_index_to_lag_grid,
+        ):
+            """Interpolate an Eulerian field onto a Lagrangian field.
+
+            Inputs:
+            the nearest_eul_grid_index_to_lag_grid(grid_dim, num_lag_nodes) and
+            interpolation weights interp_weights of
+            shape (2 * interp_kernel_width, 2 * interp_kernel_width, num_lag_nodes)
+
+            """
+            num_lag_nodes = lag_grid_field.shape[-1]
+            for i in range(0, num_lag_nodes):
+                lag_grid_field[i] = np.sum(
+                    eul_grid_field[
+                        nearest_eul_grid_index_to_lag_grid[1, i]
+                        - interp_kernel_width
+                        + 1 : nearest_eul_grid_index_to_lag_grid[1, i]
+                        + interp_kernel_width
+                        + 1,
+                        nearest_eul_grid_index_to_lag_grid[0, i]
+                        - interp_kernel_width
+                        + 1 : nearest_eul_grid_index_to_lag_grid[0, i]
+                        + interp_kernel_width
+                        + 1,
+                    ]
+                    * interp_weights[..., i]
+                ) * (dx**grid_dim)
+
         return eulerian_to_lagrangian_grid_interpolation_kernel_2d
+
     else:
+
+        @njit(cache=True, fastmath=True)
+        def vector_field_eulerian_to_lagrangian_grid_interpolation_kernel_2d(
+            lag_grid_field,
+            eul_grid_field,
+            interp_weights,
+            nearest_eul_grid_index_to_lag_grid,
+        ):
+            """Interpolate an Eulerian vector field onto a Lagrangian vector field.
+
+            Inputs:
+            the nearest_eul_grid_index_to_lag_grid(grid_dim, num_lag_nodes) and
+            interpolation weights interp_weights of
+            shape (2 * interp_kernel_width, 2 * interp_kernel_width, num_lag_nodes)
+
+            """
+            num_lag_nodes = lag_grid_field.shape[-1]
+            for i in range(0, num_lag_nodes):
+                # numba doesnt allow multiple axes for np.sum :/,
+                # hence needs to be done serially
+                lag_grid_field[0, i] = np.sum(
+                    eul_grid_field[
+                        0,
+                        nearest_eul_grid_index_to_lag_grid[1, i]
+                        - interp_kernel_width
+                        + 1 : nearest_eul_grid_index_to_lag_grid[1, i]
+                        + interp_kernel_width
+                        + 1,
+                        nearest_eul_grid_index_to_lag_grid[0, i]
+                        - interp_kernel_width
+                        + 1 : nearest_eul_grid_index_to_lag_grid[0, i]
+                        + interp_kernel_width
+                        + 1,
+                    ]
+                    * interp_weights[..., i]
+                ) * (dx**grid_dim)
+                lag_grid_field[1, i] = np.sum(
+                    eul_grid_field[
+                        1,
+                        nearest_eul_grid_index_to_lag_grid[1, i]
+                        - interp_kernel_width
+                        + 1 : nearest_eul_grid_index_to_lag_grid[1, i]
+                        + interp_kernel_width
+                        + 1,
+                        nearest_eul_grid_index_to_lag_grid[0, i]
+                        - interp_kernel_width
+                        + 1 : nearest_eul_grid_index_to_lag_grid[0, i]
+                        + interp_kernel_width
+                        + 1,
+                    ]
+                    * interp_weights[..., i]
+                ) * (dx**grid_dim)
+
         return vector_field_eulerian_to_lagrangian_grid_interpolation_kernel_2d