[WIP] Fixes #101 

I suspect it might be coming from numba. This PR is opened prior to completely fixing the issue so that the github actions get executed with each attempt to debug the underlying issue.