@fankiat 2022-12-01
It seems that `fastmath` in `numba` is doing something weird in `generate_eulerian_to_lagrangian_grid_interpolation_kernel_2d`. Removing `fastmath` seems to pass the CI consistently so far. This again could be system dependent, since both with and without `fastmath` works fine on my MBA 2022 (ARM M2) laptop; I will test this again when I get my hands on my MBP 2018 (Intel). In any case, removing `fastmath` does not seem to change the runtime much. I've measured some preliminary timings and below are the comparisons averaged over 1000 kernel calls.

(1) With `fastmath`
```
Scalar eul-to-lag grid interp took 0.027365291956812143 s/call
Vector eul-to-lag grid interp took 0.05446633300743997 s/call
```

(2) Without `fastmath`
```
Scalar eul-to-lag grid interp took 0.029281291994266212 s/call
Vector eul-to-lag grid interp took 0.06566937500610948 s/call
```

I will perform a few more tests before proceeding with removing the `fastmath` option from `numba`.