@fankiat 2022-12-01
I did a quick and dirty profiling for the kernels in `EulerianLagrangianGridCommunicator` with and without `fastmath`, for the same settings as above (`mpirun -np 4`, eulerian grid size `(2048, 2048)` and lagrangian grid size `(3, 1000)`) with timings averaged over 1000 kernel calls. The results are as below, which suggests little benefit of having `fastmath`. Therefore, for the time being, until a proper profiling tool is available and an extensive performance evaluation is done, I am removing all the `fastmath` features in the kernels using `numba`.

With `fastmath`
```
local_eulerian_grid_support_of_lagrangian_grid_kernel took 8.015662501566112e-05 s/call
eulerian_to_lagrangian_grid_interpolation_kernel (scalar) took 2.886795892845839e-05 s/call
eulerian_to_lagrangian_grid_interpolation_kernel (vector) took 6.536329200025648e-05 s/call
lagrangian_to_eulerian_grid_interpolation_kernel (scalar) took 0.0002470621250104159 s/call
lagrangian_to_eulerian_grid_interpolation_kernel (vector) took 0.00039066899998579176 s/call
interpolation_weights_kernel took 0.00016611195809673518 s/call
```

Without `fastmath`
```
local_eulerian_grid_support_of_lagrangian_grid_kernel took 7.912083296105266e-05 s/call
eulerian_to_lagrangian_grid_interpolation_kernel (scalar) took 3.2219332992099224e-05 s/call
eulerian_to_lagrangian_grid_interpolation_kernel (vector) took 5.795166699681431e-05 s/call
lagrangian_to_eulerian_grid_interpolation_kernel (scalar) took 0.0001622442079242319 s/call
lagrangian_to_eulerian_grid_interpolation_kernel (vector) took 0.00030771862505935135 s/call
interpolation_weights_kernel took 0.00016169204202014952 s/call
```