@bhosale2 2022-12-02
@fankiat a minor comment here, once you figure out the settings that work, I would recommend one of the two options:

1. Rebase the commit history to that final commit, to keep the commit history clean.
2. Make a new PR to resolve this, with that commit as the only commit. This PR can then be closed without merge, and kept in the log for future reference.