@fankiat 2022-12-05
@bhosale2 some updates here. As I work on this, I am starting to think this is potentially a combined issue from both `numba` (likely caching) and the possible existence of [flaky test](https://docs.pytest.org/en/7.1.x/explanation/flaky.html) (typically test suites running in parallel are more susceptible to). In order to mitigate both potential sources of issue, I have done the following and it seems to have resolved the issue (github actions passed ~10 times without any issue, see [4bb81eb](https://github.com/fankiat/sopht-mpi/pull/107/commits/4bb81ebf519c52fb877076deac91fb52ebe915ed)).

## For mitigating possible issues from `numba`
- [ ]  Remove `cache=True` for all `numba` kernels. The rationale here is `sopht-mpi` is meant to be executed with a larger number of processors and it is probably better for all the processors to spend some time compiling their own `numba` kernels than to load from disk (see [issue](https://github.com/numba/numba/issues/4549) mentioned by one of the main contributor of `numba`). This makes sense especially given that we don't know if `numba` is tested to be robust in cache read/write on a decently large scale (again mentioned [here](https://github.com/numba/numba/issues/3470#issuecomment-435962421), by the same contributor)
- [ ] Rearrange the order of compiling eul-lag interpolation kernels when it can either be a scalar or vector kernel. The updated `EulerianLagrangianGridCommunicator2D.py` ensures that only required kernels are compiled and returned (i.e. moving kernel generation into the `if-else` branch for checking `n_components`, instead of generating both scalar and vector versions of the kernels and returning only what is requested)

## For mitigating possible issues coming from `pytest`

- [ ]  Introduce [pytest cache clearing](https://docs.pytest.org/en/7.1.x/how-to/cache.html#clearing-cache-content), as recommended for invocations from Continuous Integration servers where isolation and correctness are more important than speed.
- [ ]  Update the test approach in immersed boundary test cases by first reducing the results from all ranks before asserting correctness. The result of doing this will cause all ranks to fail a test if any **one* rank performing the same test failed. The goal of the updated approach is to open up options for mitigating approaches like rerunning failed (flaky) tests, as suggested [here](https://docs.pytest.org/en/7.1.x/explanation/flaky.html#plugins), without running into notorious deadlock situations.
- [ ] Allow for reruns via [`pytest-rerunfailes`](https://github.com/pytest-dev/pytest-rerunfailures) when flaky tests result occurs.