@fankiat 2022-12-06
Some additional notes on this: There are two errors that typically cause the CI to fail.

1. Broadcast error, where the arrays in `numba` kernels are finding inconsistency in the shapes of the arrays it is working with. This I believe is the error coming from `numba` caching when multiple processes are involved. The changes in this PR to `numba` kernels should resolve that.
2. Value error, where sometimes interpolated values on the lagrangian grid have one element that is slightly different from the solution we draw from `sopht`. This issue is a little bit different, I am starting to think that this is coming from the fact that the test case is set up assuming zero rounding errors. I will look into this probably in a separate issue. In any case, the changes from the `pytest` side in this PR nonetheless provide an additional safeguard against any `pytest` caching issues and flaky tests.

Please let me know if you want to discuss this further to clarify things. Thanks @bhosale2 !